update updateDisplay and runEquation to descriptive function names
